Fix compilation error, add MapperTest
We split the mapper tests into two part: basic and advanced. For basic
mapper test, we just use the mapper the generate some LLVM IR or PTX IR
instruction and verify the output is correct (by our eye). For advanced
mapper test, we would need to insert the generated PTX IR into another
skeleton (generated from CUDA code with .pivot intrinsics so that we
know there to insert the instruction)

	Testing done: No
	Reviewer: N/A